added docstrings and typescripting for SBML components
