+++ sbml4humans/src/components/layout/HomeInfo.vue
@@ -15,6 +15,8 @@
 </template>
 
 <script lang="ts">
+import { defineComponent } from "@vue/runtime-core";
+
 /* Components */
 import ExamplesListContainer from "@/components/layout/ExamplesListContainer.vue";
 import FileUploadForm from "@/components/layout/FileUploadForm.vue";
@@ -26,13 +28,13 @@ import About from "@/components/layout/About.vue";
  *      - List of Examples
  *      - About Info
  */
-export default {
+export default defineComponent({
     components: {
         "examples-list": ExamplesListContainer,
         "upload-sbml": FileUploadForm,
         about: About,
     },
-};
+});
 </script>
 
 <style scoped lang="scss">
