+++ sbml4humans/src/components/sbml/Constraint.vue
@@ -10,18 +10,21 @@
     </div>
 </template>
 
-<script>
+<script lang="ts">
 import TYPES from "@/sbmlComponents";
+import { defineComponent } from "@vue/runtime-core";
 
-export default {
+/**
+ * Component to define display of Constraint objects.
+ */
+export default defineComponent({
     props: {
-        info: TYPES.Constraint,
+        info: {
+            type: Object,
+            default: TYPES.Constraint,
+        },
     },
-
-    data() {
-        return {};
-    },
-};
+});
 </script>
 
 <style lang="scss" scoped>
