+++ sbml4humans/src/components/sbml/GeneProduct.vue
@@ -12,18 +12,21 @@
     </div>
 </template>
 
-<script>
+<script lang="ts">
 import TYPES from "@/sbmlComponents";
+import { defineComponent } from "@vue/runtime-core";
 
-export default {
+/**
+ * Component to define display of GeneProduct objects.
+ */
+export default defineComponent({
     props: {
-        info: TYPES.GeneProduct,
+        info: {
+            type: Object,
+            default: TYPES.GeneProduct,
+        },
     },
-
-    data() {
-        return {};
-    },
-};
+});
 </script>
 
 <style lang="scss" scoped>
