+++ sbml4humans/src/components/sbml/Port.vue
@@ -32,16 +32,19 @@
 
 <script>
 import TYPES from "@/sbmlComponents";
+import { defineComponent } from "@vue/runtime-core";
 
-export default {
+/**
+ * Component to define display of Port objects.
+ */
+export default defineComponent({
     props: {
-        info: TYPES.Model,
+        info: {
+            type: Object,
+            default: TYPES.Port,
+        },
     },
-
-    data() {
-        return {};
-    },
-};
+});
 </script>
 
 <style lang="scss" scoped>
